basic load/save seems to work now
the store() and especially restore() routines look quite messy because of all the error checking.
It would be great to put some expectations into the CSMOD::Io class to let it check for errors.
Since everything is based on XML, it should not be a problem to change the io backend at any time.